Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-grid/layout-algorithm/flex-and-intrinsic-sizes-001-expected.txt

XXX XXX
XXX XXX
XXX XXX
XXX XXX
XXX XXX
XXX XXXX XX X XX XXX
XXX XXXX XX X XX XXX
XXXX
XXXX XX XX
XXX XX X
XXXX XXX XXXX XXX XXXX

PASS .grid 1
PASS .grid 2
FAIL .grid 3 assert_equals:
<div class="container">
  <div class="grid min-content" data-expected-width="30" data-expected-height="20" style="grid-template-columns:minmax(0, 1fr);">
    <div>XXX XXX</div>
  </div>
</div>
width expected 30 but got 0
PASS .grid 4
PASS .grid 5
PASS .grid 6
PASS .grid 7
PASS .grid 8
